Final modifications
