+++ setup/envs/graphviz.yml
@@ -1,0 +1,96 @@
+name: graphviz
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - _openmp_mutex=5.1=1_gnu
+  - atk-1.0=2.36.0=h3371d22_4
+  - blas=1.0=openblas
+  - bzip2=1.0.8=h7b6447c_0
+  - ca-certificates=2023.08.22=h06a4308_0
+  - cairo=1.16.0=hf32fb01_1
+  - cffi=1.15.1=py310h5eee18b_3
+  - expat=2.4.8=h27087fc_0
+  - filelock=3.9.0=py310h06a4308_0
+  - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
+  - font-ttf-inconsolata=3.000=h77eed37_0
+  - font-ttf-source-code-pro=2.038=h77eed37_0
+  - font-ttf-ubuntu=0.83=hab24e00_0
+  - fontconfig=2.13.1=h6c09931_0
+  - fonts-conda-ecosystem=1=0
+  - fonts-conda-forge=1=0
+  - freetype=2.10.4=h0708190_1
+  - fribidi=1.0.10=h36c2ea0_0
+  - gdk-pixbuf=2.42.6=h04a7f16_0
+  - gettext=0.19.8.1=h73d1719_1008
+  - giflib=5.2.1=h36c2ea0_2
+  - glib=2.68.4=h9c3ff4c_1
+  - glib-tools=2.68.4=h9c3ff4c_1
+  - gmp=6.2.1=h295c915_3
+  - gmpy2=2.1.2=py310heeb90bb_0
+  - graphite2=1.3.13=h58526e2_1001
+  - graphviz=2.49.0=h85b4f2f_0
+  - gtk2=2.24.33=h539f30e_1
+  - gts=0.7.6=h64030ff_2
+  - harfbuzz=3.0.0=h83ec7ef_1
+  - icu=68.2=h9c3ff4c_0
+  - jbig=2.1=h7f98852_2003
+  - jinja2=3.1.2=py310h06a4308_0
+  - jpeg=9e=h166bdaf_1
+  - ld_impl_linux-64=2.38=h1181459_1
+  - lerc=2.2.1=h9c3ff4c_0
+  - libdeflate=1.7=h7f98852_5
+  - libffi=3.4.4=h6a678d5_0
+  - libgcc-ng=11.2.0=h1234567_1
+  - libgd=2.3.3=h6ad9fb6_0
+  - libgfortran-ng=11.2.0=h00389a5_1
+  - libgfortran5=11.2.0=h1234567_1
+  - libglib=2.68.4=h174f98d_1
+  - libgomp=11.2.0=h1234567_1
+  - libiconv=1.17=h166bdaf_0
+  - libopenblas=0.3.21=h043d6bf_0
+  - libpng=1.6.37=h21135ba_2
+  - libprotobuf=3.20.3=he621ea3_0
+  - librsvg=2.52.2=hc3c00ef_0
+  - libstdcxx-ng=11.2.0=h1234567_1
+  - libtiff=4.3.0=hf544144_1
+  - libtool=2.4.6=h9c3ff4c_1008
+  - libuuid=1.41.5=h5eee18b_0
+  - libwebp=1.2.2=h3452ae3_0
+  - libwebp-base=1.2.2=h7f98852_1
+  - libxcb=1.15=h7f8727e_0
+  - libxml2=2.9.12=h72842e0_0
+  - lz4-c=1.9.3=h9c3ff4c_1
+  - markupsafe=2.1.1=py310h7f8727e_0
+  - mpc=1.1.0=h10f8cd9_1
+  - mpfr=4.0.2=hb69a4c5_1
+  - mpmath=1.3.0=py310h06a4308_0
+  - ncurses=6.4=h6a678d5_0
+  - networkx=3.1=py310h06a4308_0
+  - ninja=1.10.2=h06a4308_5
+  - ninja-base=1.10.2=hd09550d_5
+  - numpy=1.26.0=py310heeff2f4_0
+  - numpy-base=1.26.0=py310h8a23956_0
+  - openssl=3.0.11=h7f8727e_2
+  - pango=1.48.10=h54213e6_2
+  - pcre=8.45=h9c3ff4c_0
+  - pip=23.2.1=py310h06a4308_0
+  - pixman=0.40.0=h36c2ea0_0
+  - pycparser=2.21=pyhd3eb1b0_0
+  - python=3.10.13=h955ad1f_0
+  - python-graphviz=0.20.1=pyh22cad53_0
+  - pytorch=2.0.1=cpu_py310hab5cca8_0
+  - readline=8.2=h5eee18b_0
+  - setuptools=68.0.0=py310h06a4308_0
+  - sqlite=3.41.2=h5eee18b_0
+  - sympy=1.11.1=py310h06a4308_0
+  - tk=8.6.12=h1ccaba5_0
+  - torchview=0.2.6=pyhd8ed1ab_0
+  - typing-extensions=4.7.1=py310h06a4308_0
+  - typing_extensions=4.7.1=py310h06a4308_0
+  - tzdata=2023c=h04d1e81_0
+  - wheel=0.41.2=py310h06a4308_0
+  - xz=5.4.2=h5eee18b_0
+  - zlib=1.2.13=h5eee18b_0
+  - zstd=1.5.0=ha95c52a_0
